backend/requirements.txt
asgiref==3.2.3
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Django==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
flickrapi==2.4.0
idna==2.8
oauthlib==3.1.0
psycopg2==2.8.4
pycodestyle==2.5.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
